All get routes now follow a MVC format
