+++ server/controllers/registrationRoutes.ts
@@ -1,6 +1,4 @@
-const { User } = require('../model/user-database');
 const registrationController = require('../controllers/registrationController');
-import { Req, Res, QueryString } from '../utils/types-routes';
 
 var express = require('express');
 const registrationRoutes = express.Router();
