+++ server/utils/static-middleware.ts
@@ -1,5 +1,5 @@
 const path = require('path');
-const express = require('express');
+var express = require('express');
 
 const publicPath = path.join(__dirname, '../../public');
 const staticMiddleware = express.static(publicPath);
